don't try to drop poitns column

+++ database/migrations/2024_12_26_210000_remove_points_from_users_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('points');
-        });
+        if (Schema::hasColumn('users', 'points')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('points');
+            });
+        }
     }
 
     /**
